+++ src/main/java/com/universe/uni/dto/RoundMissionDto.java
@@ -1,24 +1,24 @@
-package com.universe.uni.dto.response;
+package com.universe.uni.dto;
 
 import java.time.LocalDateTime;
 
 import com.universe.uni.domain.GameResult;
-import com.universe.uni.domain.entity.MissionContent;
 import com.universe.uni.domain.entity.RoundMission;
+import com.universe.uni.dto.response.RoundMissionContentDto;
 
 import lombok.Getter;
 
 @Getter
 public class RoundMissionDto {
 	private long id;
-	private MissionContentDto missionContent;
+	private RoundMissionContentDto missionContent;
 	private GameResult result;
 	private GameResult finalResult;
 	private LocalDateTime updatedAt;
 
 	public RoundMissionDto(RoundMission roundMission) {
 		this.id = roundMission.getId();
-		this.missionContent = new MissionContentDto(roundMission.getMissionContent());
+		this.missionContent = new RoundMissionContentDto(roundMission.getMissionContent());
 		this.result = roundMission.getResult();
 		this.finalResult = roundMission.getFinalResult();
 		this.updatedAt = roundMission.getUpdatedAt();
