+++ src/main/java/com/universe/uni/dto/response/GameReportResponseDto.java
@@ -1,6 +1,7 @@
 package com.universe.uni.dto.response;
 
 import com.universe.uni.domain.entity.RoundMission;
+import com.universe.uni.dto.RoundMissionDto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
