+++ src/main/java/com/universe/uni/dto/response/RoundMissionContentDto.java
@@ -5,13 +5,13 @@ import com.universe.uni.domain.entity.MissionContent;
 import lombok.Getter;
 
 @Getter
-public class MissionContentDto {
+public class RoundMissionContentDto {
 	private Long id;
 	private MissionCategoryDto missionCategory;
 	private String content;
 	private String image;
 
-	public MissionContentDto(MissionContent missionContent) {
+	public RoundMissionContentDto(MissionContent missionContent) {
 		this.id = missionContent.getId();
 		this.missionCategory = new MissionCategoryDto(missionContent.getMissionCategory());
 		this.content = missionContent.getContent();
